+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## **(V.4.5.3 Changes) (1.21.10 Minecraft)**
+## **(V.4.5.3 Changes) (1.21.11 Minecraft)**
 
 ##### Major:
-Updated to 1.21.10
+Updated to 1.21.11
+
+Note, this uses gizmos to render the outlines. But gizmos have a bug in minecraft where translucent blocks show through the gizmo.
+Vanilla minecraft will have to fix this on their end.